refactor signout strategy
